==> 01_download_preprocess/02_Merge_single_files.ipynb <==
import pandas as pd
import glob
import os

# Content

This script reads all .csv files and export the relevant columns. It combines the .csv files to a single file

# Setting parameters

# Folder with the csv files
PATH = '../data/02_processed_csvs/*.csv'
# Selected columns to export
COL_EXPORT_MANUAL = ['file', 
                     'decision_date', 
                     'jurisdiction_id', 
                     'jurisdiction_name', 
                     'jurisdiction_name_long', 
                     'court_name', 
                     'court_name_abbreviation',
                     'cb_status',
                     'cb_no_attorneys',
                     'cb_no_parties',
                     'cb_no_opinions',
                     'cb_no_judges',
                     'citation1_type',
                     'reporter_full_name',
                     'cb_data_attr1_first_attorneys',
                     'cb_data_judge1',
                     'cb_data_party1']

# Process data

files = glob.glob(PATH)

df = pd.read_csv(files[0])
df['file'] = files[0]

# Add all category columns (starting with cb_data_text_0)
col_export = COL_EXPORT_MANUAL + [x for x in list(df.columns) if 'cb_data_text_0' in x]

# Iterate over all csv files
df = df[col_export]
for file in files[1:len(files)]:
    print(file)
    df_tmp = pd.read_csv(file)
    df_tmp['file'] = file
    df_tmp = df_tmp[col_export]
    df = df.append(df_tmp)

# Save the data

df.shape

df.to_csv('../data/03_combined/df_combined_v2.csv.gz', compression='gzip')
